LPS-66685 Consistency
